Moved Utils form private to public
