+++ index.html
@@ -129,6 +129,7 @@
         </div>
     </div>
     <canvas id ="canvas" width="500", height="500"></canvas>
+    <script src="./models/Utils.js"></script>
     <script src="./models/Img.js"></script>
     <script src="./models/GameObject.js"></script>
     <script src="./models/Collectible.js"></script>
